[RTC] Experience Fragment Component with Localization Logic #558
This is a new implementation. The former one was based on two configuration properties (localization root and depth) to deduct the localization structure. This implementation automatically detects the localization structure of the site by relying on the LanguageManager and the LiveRelationshipManager APIs.
The logic is only triggered for XF resources defined in a template.