Issue 703: [Navigation] Prevent NPE when navigation root missing jcr:content
Prevents a null pointer exception when the specified navigation root page
is missing it's jcr:content child resource.

In this case, the language root for the navigation root is determined
by using the resource located at the navigation root page path instead
of the jcr:content resource.

The root page is alerady asserted to exist; therefore, the resource
at the same path should resolve.