Fix possible DataLayer NPEs (#1029)
* Add NotNull annotation to AbstractContainerImpl#getItems()
* Add NotNull annotation to PanelContainerImpl#readItems()
* Fix possible NPE in PanelContainerImpl#getItemModels()

Optional#get() was performed without #isPresent() check, followed
by a null check on the result of the #get() call.

Optional#get() does not return null if the optional is empty, it
causes an exception.

The Optional#get() call is replaced with Optional#ifPresent()
instead.

* Remove useless null check in AccordionImpl#getDataLayerShownItems()

AccordionImpl#getItems() is annotated NotNull, a null check on its
return value is not needed and is removed.

* Simplify collecting of expandedItemIds in AccordionImpl

Replaces the collection of a stream -> list followed by converting
the list -> array. The stream can be converted directly to an array.

* Fix possible NPE in AccordionImpl#getDataLayerShownItems()

Resolves possible NPE resulting from Component.getData().getId()
where Component.getData() may return null.

Null values are filtered before getting the id.

* Simplify TabsImpl#getDataLayerShownItems()

Removes several placeholder variables that are not needed.
Streams/Optionals are already in use, they can accomplish everything
that needs to be done in this method.

* Fix possible NPE in PanelContainerItemImpl constructor

Fixes a possible NPE in the constructor for
PanelContainerItemImpl. The second parameter in a
ValueMap#get(property, default) should never be null.

The default value is replaces with the class name and
the result is wrapped in an Optional, allowing the
default to only be evaluated if required.