Fix NPE in container implementation test
The Tabs, Accordion, and Carousel container model
implementations require the resource to be non-nul.

In the tests, the models should not be constructed
directly without using the model factory. Unless
constructor injection is used, models can not be
fairly initialized in this manner.

empty container tests are updated to initialize the
model properly such that the resource is not null.