Make AbstractContainerImpl#readItems() abstract
There are two implementations of this method (one returning
ResourceListItems and the other returning PanelListItems), the
default method implementation is removed from AbstractContainerImpl
and moved to the implementation class where it is actually used
instead of implementing and then overriding. This forces container
implementations to be cognisant of what items they are containing
and makes the refactoring of AbstractContainerImpl less likely to
introduce bugs.

Additionally, most of the AbstractContainerImpl tests are removed,
the tests of the items returned belong in the final container
implementation tests, not the abstract container.