Simplify and improve performance of ListItem
Reworks ListImpl to eliminate "populate" functions that
mutate the list items, followed (potentially) by other
functions such as sorting and limiting, which again mutate
the list items. This pattern makes the state of the list
items field difficult to track.

All of the "populate" functions are refactored to getters
that return a stream of results. There is now a single
location where the list items are set once, and are not
subsequently modified.

The refactoring of the source selector switch will facilitate
moving to enhanced switch expression in future JDK versions.

The sorting of the results is done in the stream returned
by the appropriate "populate" function (now a getter).

As the result streams are (in most cases) lazy-loading, the
limiting to the maximum number of results is performed on
the result stream and can prevent the delving of unnecessary
data. This improves performance when there are a large number
of candidate results that will be trimmed due to exceeding the
desired maximum number of list items.