Update list consistency tests to verify results and ordering
Tests of list results are updated to specify the result items and
order instead of just testing the result length. This protects
against future changes that may inadvertently change the list result
order, or (somehow) replace results with unintentional results such
that the length remains the same and the content differs.