Issue/1045: Simplify and improve ListImpl performance (#1046)
* Remove useless "throws Exception" from ListItem tests

Removes "throws Exception" from method signature of methods
that do not throw a checked exception.

* Use `assertArrayEquals` when comparing arrays in ListImplTest

Consistency checks are siplified to use `assertArrayEquals` which
performs a length check, and is the standard way to test arrays
for equality.

Also renames `checkListConsistency` -> `checkListConsistencyByTitle`
so as to not be confused with `checkListConsistencyByPath`.

* Update list consistency tests to verify results and ordering

Tests of list results are updated to specify the result items and
order instead of just testing the result length. This protects
against future changes that may inadvertently change the list result
order, or (somehow) replace results with unintentional results such
that the length remains the same and the content differs.

* Remove use of deprecated `org.mockito.Matchers.any`

`org.mockito.Matchers.any` is deprecated and replaced with
`org.mockito.ArgumentMatchers.any`.

* Simplify and improve performance of ListItem

Reworks ListImpl to eliminate "populate" functions that
mutate the list items, followed (potentially) by other
functions such as sorting and limiting, which again mutate
the list items. This pattern makes the state of the list
items field difficult to track.

All of the "populate" functions are refactored to getters
that return a stream of results. There is now a single
location where the list items are set once, and are not
subsequently modified.

The refactoring of the source selector switch will facilitate
moving to enhanced switch expression in future JDK versions.

The sorting of the results is done in the stream returned
by the appropriate "populate" function (now a getter).

As the result streams are (in most cases) lazy-loading, the
limiting to the maximum number of results is performed on
the result stream and can prevent the delving of unnecessary
data. This improves performance when there are a large number
of candidate results that will be trimmed due to exceeding the
desired maximum number of list items.

* Rename ListImpl.getListItems(Source listType)

Renames `ListImpl.getListItems(Source listType)` to
`ListImpl.getPages()` to avoid confusion with `List.getListItems()`.

It is unnecessary to pass the list type, as the list type must not
be anything other than provided in `ListImpl.getListType()`.