Issue/1030: Annotate optional sling model fields as `@Nullable` and fix potential NPEs (#1040)
* Add @Nullable annotation to optional injected fields

Adds the @Nullable annotation to any injected model field
that is marked as optional and does not provide a default
value.

These fields may be left null by the Sling model factory
and the possibility of null must be handled accordingly.

* Make AbstractContainerImpl#readItems() abstract

There are two implementations of this method (one returning
ResourceListItems and the other returning PanelListItems), the
default method implementation is removed from AbstractContainerImpl
and moved to the implementation class where it is actually used
instead of implementing and then overriding. This forces container
implementations to be cognisant of what items they are containing
and makes the refactoring of AbstractContainerImpl less likely to
introduce bugs.

Additionally, most of the AbstractContainerImpl tests are removed,
the tests of the items returned belong in the final container
implementation tests, not the abstract container.

* Correct potential NPE in AbstractContainerImpl

Corrects possible NPE exceptions that may arise from
`currentStyle` being null.

Also, reworks style initializers to not rely on side effects.

* Simplify AccordionImpl#getExpandedItems()

Series of nested loops and conditionals is simplified to
a sequence of optional/stream filter and maps.

* Correct potential NPE in CarouselImpl

Corrects possible NPE exceptions that may arise from
`currentStyle` being null.

* Correct potential NPE in V2 PageImpl

Corrects possible NPE exceptions that may arise from
`currentStyle` being null.

* Fix NPE in container implementation test

The Tabs, Accordion, and Carousel container model
implementations require the resource to be non-nul.

In the tests, the models should not be constructed
directly without using the model factory. Unless
constructor injection is used, models can not be
fairly initialized in this manner.

empty container tests are updated to initialize the
model properly such that the resource is not null.

* Correct potential NPE in TabImpl

Corrects a potential NPE in TabImpl due to
`activeItem` possibly being null. The Sling
API declares that for Resource.getChild(name),
name must not be null.

* Remove unused injected field in ButtonImpl

The `resource` field is not referenced
in the model, it does not need to be injected.

* Correct potential NPE in ClientLibraryImpl

The ClientLibraryImpl specifies the default
injection strategy as OPTIONAL; thus all
injected fields may be null and must perform
null checks.

* Correct potential NPE in V1 image

Corrects a potential NPE in V1 Image
model implementation that can arise when
`fileReference` is null.

The Sling API specifies that for
ResourceResolver.getResource(name), name
must not be null.

* Add NotNull annotation to ImageResourceWrapper#getResourceType()

This method overrides a method that is annotated with @NotNull;
therefore, this method should be annotated as well.

* Add NotNull annotations to CoreFormHandlingServlet

Adds the @NotNull to the servlet doPost method parameters
as defined by SlingAllMethodsServlet.
Also, removes some unnecessary fields.

* Mark compareTo() parameter as @NotNull

Specification for Comparable.compareTo() indicates the
expected result of compareTo(null) is a NPE.

* Remove null check on resource.getValueMap()

Resource#getValueMap() may not return null. The null check
is unnecessary. The nested if statements can be removed and
simplified.