Remove `ComponentDataImpl` & replace with `DataLayerBuilder`
Removes the `ComponentDataImpl` class, replaces it's use in
`AbstractComponentImpl` with the `DataLayerBuilder`.

Eliminates all of the `getDataLayer*()` methods in the abstract
component implementation, and instead requires that each component
model that overrides the data layer provides it's own implementation
of `getComponentData()` - which can use `DataLayerBuilder.extending()`
to inherit data from a super model.