Make `PageListItemImpl#getRedirectTarget` package-private & static
Updates PageListItemImpl to only determine the redirect target
if shadowing is enabled. This has performance benefits because
there is no reason to ever compute the redirect target if it
will not be used.

The #getRedirectTarget() method is made static because it does not
reference any class fields and is effectively static.

The #getRedirectTarget() method is made package-private so that
the logic can be re-used.