Split `NavigationImpl#getItems()` into `getItems()` + `getNavigationRoot()`
Simplifies the logic for getting the navigation items by splitting
off the work of determining the effective navigation root into it's
own method.

The inner class `NavigationImpl.NavigationRoot` is eliminated because it
is no longer needed. The depth awareness for a navigation item is derived
using the page depth of the item vs the page depth of the effective
navigation root + structureStart.