Add ability to extend existing Component/AssetData
Adds data layer suppliers that are backed by existing
data layer models. This provides a mechanism by which
an existing ComponentData or AssetData can be used to
initialize a builder and allow one or more fields to
be changed without the work of re-supplying every field
value.

The intended use-case for this feature is when one
Component extends another, so too can the ComponentData
extend the original Components data.