[Image] Allow lazy threshold to be set in design policy #1090 (#1091)
* Add lazy threshold property and getter to Image model interface

Adds `Image#getLazyThreshold()` and `Image#PN_DESIGN_LAZY_THRESHOLD`
to the Image interface.

* Add lazy threshold getter to V2 Image model implementation

Implements `Image#getLazyThreshold()` in the V2 Image model
implementation.

* Update V2 Image tests for lazy threshold

Add test for setting lazy threshold via design policy.
Update expected exporter JSON to include the lazy threshold
property.

* Add lazy threshold data attribute to V2 image HTL

Adds the `data-cmp-lazythreshold` attribute to the V2
image element dataset, so that this value can be used
by the JS to determine when the image should load.

* Use lazy threshold data attribute in V2 image JS

Updates the V2 image JS to use the lazy threshold defined
in the image element dataset instead of the constant
`LAZY_THRESHOLD = 0`.

* Add lazy threshold to V2 image design dialog

Adds the lazy threshold property to the design dialog
for the v2 image. The threshold may be set to any
non-negative number.

* Remove `uuidDisabled` member from ImageImpl

The `uuidDisabled` field is never referenced outside of the `initModel()`
method, it is moved to that scope and not retained after `initModel()`
has completed.

* Add documentation to member fields/methods in V2 ImageImpl

For clarity, and future developers, doc comments are added to any
undocumented class members.

* Add documentation to member fields/methods in ImageAreaImpl

For clarity, and future developers, doc comments are added to any
undocumented class members.

Update readme for image component

- add new design dialog property for lazyThreshold feature.