[Navigation] Fix inconsistent `isActive` behaviour caused by redirect chains + cleanup (#1080)
* Make `PageListItemImpl#getRedirectTarget` package-private & static

Updates PageListItemImpl to only determine the redirect target
if shadowing is enabled. This has performance benefits because
there is no reason to ever compute the redirect target if it
will not be used.

The #getRedirectTarget() method is made static because it does not
reference any class fields and is effectively static.

The #getRedirectTarget() method is made package-private so that
the logic can be re-used.

* Add comments to private fields/methods in NavigationImpl

Adds comments to many private fields and methods in the
`NavigationImpl`.

* Remove star import in NavigationImpl

Removes a star import from NavigationImpl.

* Update NavigationImpl to check more than one level of redirect

The NavigationImpl is updated to use the
`PageListItemImpl#getRedirectTarget(Page)` method to determine
if a navigation item is selected/active.

The existing implementation would check only one level of redirect
while the PageListItemImpl determines the effective item by following
an entire redirect chain.

This eliminates inconsistencies in navigation where two effectively
identical navigation items (same URL/title) may have different active
states based on the redirect chain used to reach those final items.

* Remove injected ResourceResolver from NavigationImpl

The injected ResourceResolver is removed; the resource
resolver can be accessed via the resource or request.

* Remove injected properties/accessibility label from NavigationImpl

Removes the injection of the properties and the optional
accessibility label property. The properties are already
available via the resource.

* Move `NavigationImpl#skipNavigationRoot` field -> initModel() method

The skipNavigationRoot flag is only used in the initializer, there
is no need to maintain it after the initializer method has returned.

* Simplify `NavigationImpl#getRootItems()`

Replaces triple-nested while loops with a recursive
call to build a stream of navigation root items.

* Split `NavigationImpl#getItems()` into `getItems()` + `getNavigationRoot()`

Simplifies the logic for getting the navigation items by splitting
off the work of determining the effective navigation root into it's
own method.

The inner class `NavigationImpl.NavigationRoot` is eliminated because it
is no longer needed. The depth awareness for a navigation item is derived
using the page depth of the item vs the page depth of the effective
navigation root + structureStart.

* Only fetch live relationships if needed in NavigationImpl

The live relationships are used conditionally,
the call to the relationship manager is moved
so that it is only done if the results will be
used

* Add tests for asserting isActive behaviour of redirect chains

Adds tests to verify that navigation items are marked as active if
the redirect chain eventually leads to a page which is considered
active.

Fixes #1079