Fix possible NPE in LayoutContainerImpl
Fixes a possible NPE in `LayoutContainerImpl` that can occur if the
layout type is not set in the component properties or style properties
and the current resource's value map does not gracefully handle a
null default value.

Additionally, the injected `currentStyle` field is removed because
it is already injected into `AbstractContainerImpl`, which
`LayoutContainerImpl` extends.