[Teaser] Fix inconsistent `linkURL` behaviour + cleanup (#1106)
- Removes the testing of debug logs from TeaserImpl tests.
There is no value to asserting that debug messages were
logged.
- Adds tests to ensure that the teaser link URL, when
derived from the actions, does not require that the
action reference an internal page.
I.e. the teaser link behaviour should not change if
the first CTA link references an external page.
- Corrects the Teaser link URL behaviour where the URL of the first
action is not used unless it resolves to a page.
- Updated behaviour is to use the first actions link URL regardless
of if the path resolves to or a page or not.
- Simplifies the TeaserImpl logic for determining the title,
pretitle, and description by removing unessisary null sets.
- Updates Teaser model implementation to simplify the check for if
the teaser has an image or not.
- The Teaserimpl.Action fields that are set in the constructor
and never updated are marked as final.
- Simplifies the method TeaserImpl#getImage() by removing the manual
wrapping of the resource, and explicitly asking the ModelFactory
for an Image model.
- Moves the field initializers out of the initModel method
and into their respective getter methods so that the logic
for determining each field value is isolated and clear.
- Removes the optional injection policy, the current style is now
required to be injected in order to construct this model.
- The teaser action resource must never be null. Allowing it to
be null would result in a NPE.
- Adds javadocs for all un-documented fields and methods in
TeaserImpl.