Remove `ExperienceFragmentImpl#resolveName(PageManager)`
Removes the method `#resolveName(PageManager)` from the XF model
implementation, and places the logic for deriving the XF name
directly in the `#getName()` method.

The name will no longer be determined unless the `#getName()`
method is called.

This simplifies understanding of the model because it makes it
obvious how the return value of `#getName()` is populated.

Additionally, this simplifies future refactoring as the ability
to call this method is no longer dependent on having previously
called the `#resolveName(PageManager)` method.