Remove `ExperienceFragmentImpl#appendCssClassNames()`
Removes the `#appendCssClassNames()` method and consolodates the
logic for determining the css class names into the `#getCssClassNames()`
method.

Class names will now only be determined if they are requested via
`#getCssClassNames()` instead of being determined in the model
initializer. This is beneficial because `#getCssClassNames()` may
never be called.

This also makes the model more tolerant to future refactoring because
the return value of `#getCssClassNames()` no longer changes depending
on if `#appendCssClassNames()` has been called, and eliminates any
possibility of accidentally calling it twice.