Use search model for settings in searchResultServlet
Moves the logic for determining the search root page
path out of the search result servlet and into the
search component model.

Updates the search result servlet to construct the
model for the search component being used, and then
rely upon that models settings for search configuration
instead of re-computing the settings inside the
search result servlet.

This ensures that the configuration used in the
search result servlet will not differ from the
settings provided in the model, and will decouple
the search result servlet from this specific search
model implementation, allowing for delegate models
to provide different settings that will be honoured
by the search result servlet.

This change also has performance benefits because it
eliminates the re-computation of the search component
ID for every search result.