Deprecate `Container#getItems()` and `AbstractContainerImpl#readItems()`
The `readItems()` abstract method was a means by which to delegate the
constructing of container `ListItem`s to the component model.
The `ListItems()` method was the mechanism by which those `ListItem`s
were exposed.

It is no longer desired that containers should be treated as lists;
therefore, these methods are marked as deprecated.