Update List component item sorting to sort based on presentation title
Updates the `ListSort` comparator, which sorts the pages returned by
the `ListImpl`, so that when sorting by "Title" it uses the same
title that `PageListItemImpl` would use, to ensure consistency in
the sorting and presentation.

The comparator is also updated so that the title colator isn't
constructed unless sorting by title is enabled.