Fix broken `ListImplTest#testOrderByTitleWithNoTitleForOneItem()`
Fixes this broken test. This test was broken because a page with
no title was previously sorted to the end of the list; however,
since the name of the page will now be taken into account when
sorting (so as to match the presentation title), the order of these
items has changed.