[List] Correct sort ordering not matching list item title (#1269) (#1273)
* Add static method `PageListItemImpl#getTitle(page)`

Adds a static method to get the page list item title from a given
page without constructing a list item. This will allow for the
sorting of pages by title without having to construct the list
items first.

* Update List component item sorting to sort based on presentation title

Updates the `ListSort` comparator, which sorts the pages returned by
the `ListImpl`, so that when sorting by "Title" it uses the same
title that `PageListItemImpl` would use, to ensure consistency in
the sorting and presentation.

The comparator is also updated so that the title colator isn't
constructed unless sorting by title is enabled.

* Fix broken `ListImplTest#testOrderByTitleWithNoTitleForOneItem()`

Fixes this broken test. This test was broken because a page with
no title was previously sorted to the end of the list; however,
since the name of the page will now be taken into account when
sorting (so as to match the presentation title), the order of these
items has changed.

* Add JavaDocs to `PageListItemImpl`

Adds javadoc comments to previously undocumented members of the
`PageListItemImpl` class.

Fixes #1269