[ExperienceFragment] Fix incorrect exported items order (#1138)
* Correct XF test items order

Corrects the expected child items order
in the expected experience fragment model
data.

The expected item order is the order the
items exist in the JCR.

* Remove `MockSlingHttpServletRequest` from XF tests

Removes `MockSlingHttpServletRequest` from the
ExperienceFragmentImpl test class because there
is no need to construct a mock request, the mock
request provided by the test context is
sufficient.

* Make `ContentFragmentUtils#getComponentExporters()` return LinkedHashMap

Updates `ContentFragmentUtils#getComponentExporters()` so that it
returns a LinkedHashMap. The ordering of the map is important, and
the returned type should reflect that.

* Correct exported items order in XF model implementation

Corrects the order provided by `ExperienceFragmentImpl#getExportedItemsOrder()`.
This is corrected by ensuring that the `children` map is a `LinkedHashMap` instead
of a `HashMap`.

* Replace `Nonnull` with `NotNull` in XF model implementation

Replaces the use of the `Nonnull` annotation with `NotNull`
in the `ExperienceFragmentImpl` model implementation so as to be
consistent with other annotations used across the project.

* Fix javadoc issues in `ExperienceFragmentImpl`

Removes unessisary javadoc on override method. The javadoc can be
inferred from the interface. This corrects an issue wherein the
javadoc has a self-referencing link, and incorrectly identifies a
method as deprecated.

Additionally, a javadoc block is removed fr the initializer method
and replaced with a line comment.

* Remove `ExperienceFragmentImpl#resolveName(PageManager)`

Removes the method `#resolveName(PageManager)` from the XF model
implementation, and places the logic for deriving the XF name
directly in the `#getName()` method.

The name will no longer be determined unless the `#getName()`
method is called.

This simplifies understanding of the model because it makes it
obvious how the return value of `#getName()` is populated.

Additionally, this simplifies future refactoring as the ability
to call this method is no longer dependent on having previously
called the `#resolveName(PageManager)` method.

* Refactor `ExperienceFragmentImpl#initModel()` to avoid unnecessary `PageManager`

The `PageManager` is only required if `currentPage` is null.

* Remove `ExperienceFragmentImpl#appendCssClassNames()`

Removes the `#appendCssClassNames()` method and consolodates the
logic for determining the css class names into the `#getCssClassNames()`
method.

Class names will now only be determined if they are requested via
`#getCssClassNames()` instead of being determined in the model
initializer. This is beneficial because `#getCssClassNames()` may
never be called.

This also makes the model more tolerant to future refactoring because
the return value of `#getCssClassNames()` no longer changes depending
on if `#appendCssClassNames()` has been called, and eliminates any
possibility of accidentally calling it twice.

* Don't compute all child models in XF model initializer

The creation of the child models can be a heavy operation if there
are many child models or if they cause a cascade due to those model
initializers also constructing their child models.

The child models are only needed for JSON serialization, and should
only be computed if needed.

* Remove `ExperienceFragmentImpl#resolveLocalizedFragmentVariationPath()`

Moves the logic for determining the localized fragment variation
path into the getter

* Change `ExperienceFragmentImpl.modelFactory` injector to `OSGiService`

Model factory is an OSGI service and can be injected as such without
needing to work through injection providers.

* Remove injection of ResourceResolver in XF model implementation

The ResourceResolver can easily be determined from the request without
needing to inject it.

* Remove injection of `resource` in ExperienceFragmentImpl

The `resource` is available via the request and does not need
to be injected.

This also eliminates a lot of variable shadowing.

* Change default injection strategy to `REQUIRED` in XF model implementation

All injected values are required other than `currentPage`. The default
should be reverted back to `REQUIRED` with `OPTIONAL` applied only to
the `currentPage` field.

* Add `Nullable` annotations to private methods in `ExperienceFragmentImpl`

Adds `Nullable` or `NotNull` annotations to private methods in
`ExperienceFragmentImpl`. Specifically, many of the methods belonging
to the language manager or relationship manager expect the resource
to be non-null (although they are not annotated as such).

Adding these annotations reveals a potential NPE in `#getLocalizationRoot(String)`,
which is resolved.

* Add JavaDoc comments to private members of `ExperienceFragmentImpl`

Adds JavaDoc comments to all undocumented private fields and methods in
`ExperienceFragmentImpl` to clarify their purpose.

* Add XF model tests to ensure caching of values

Ensures that values that are computed on demand return the same
object when requested again.