[Search] Fix unclosed resource resolver in SearchServlet + other refactoring (#1253)
* Ensure leaking ResourceResolver is closed in SearchResultServlet

Updates the SearchResultServlet to make a best-attempt to close
the ResourceResolver created by the QueryBuilder.

Updates the SearchResultServlet tests to check that the leaking
resource resolver has been closed.

* Fix Search model tests not testing template based search

Fixes an error in the testing that causes the `testTemplateBasedSearch()`
method to not actually test search based on the template policy properties.

Additionally, adds multiple test results to ensure that search functions
properly when more than one result is present.

* Test result IDs in SearchResultServletTest

Adds tests to confirm the result IDs returned by the
SearchResultServlet are as expected.

* Test search results empty if query is too short

Adds a test to ensure that the search results are empty if the
search query string is shorter than the minimum required search
query string length.

* Add test for behaviour of search when no component specified

Adds a test to check the behaviour of the `SearchResultServlet`
when no suffix is set and the servlet is required to rely on the
default settings instead of the settings configured in a component.

* Add null annotations to SearchResultServlet

Adds `@Nullable` & `@NotNull` annotations to the SearchResultServlet
method signatures to help detect potential NPEs.

* Remove useless null check in SearchResultServlet

Removes a pointless check if PageManager is null.

* Fix potential NPE in `SearchResultServlet`

Fixes a NPE that occurs when the search resource
is null and the language manager returns null for
the language root.

* Send 404 if currentPage doesn't exist in SearchResultServlet

Changes the SearchResultServlet to send a 404 error response code
if the `currentPage` does not exist. Also moves the logic of the
`writeJson` method into the `doGet` method.

* Make `SearchImpl` extend `AbstractComponentImpl`

The search model implementation is updated to extend
`AbstractComponentImpl`, as is the case with other components.

* Use search model for settings in searchResultServlet

Moves the logic for determining the search root page
path out of the search result servlet and into the
search component model.

Updates the search result servlet to construct the
model for the search component being used, and then
rely upon that models settings for search configuration
instead of re-computing the settings inside the
search result servlet.

This ensures that the configuration used in the
search result servlet will not differ from the
settings provided in the model, and will decouple
the search result servlet from this specific search
model implementation, allowing for delegate models
to provide different settings that will be honoured
by the search result servlet.

This change also has performance benefits because it
eliminates the re-computation of the search component
ID for every search result.

* Reuse language/live-copy logic between Navigation and Search

Navigation and Search model implementations share the same logic
for resolving a localized page path. This logic is moved to a new
`LanguageUtils` utility class to eliminate duplication.

* Send 400 on invalid offset in SearchResultServlet

Corrects where a 500 error is sent if the search offset parameter
cannot be parsed to a long. Instead, a 400 (Bad Request) will be sent.

* Add javadoc comments to SearchResultServlet members

Adds JavaDoc comments to members of the SearchResultServlet class.

* Default search root should be currentPage instead of `/content`

The search model implementation defines the default search root
as `/content`; however, this is not a page and a search will not
be performed under this path by the search servlet - it will search
the current page instead.

* Add javadoc comments to members of `SearchImpl`

Adds javadoc comments to previously undocumented/private members
of `SearchImpl`.

Co-authored-by: Burkhard Pauli <pauli@adobe.com>