Specify "cq:Component" as model resource type for `ComponentImpl`
Adds the resource type "cq:Component" to the model definition for
the default Component implementation (`ComponentImpl`).

This change allows other models to indicate `Component.class` as an
adapter, and have those models take precedence based on resource type
hierarchy, without interfering with `ComponentImpl` being used as the
default fallback for any component that does not have a model with
a more specific resource type defined.

----
#1447