[Accordion] Allow single expansion to start with all closed (#2287)
Allows a single-expansion accordion to be configured to start with
all panels closed by default. Previously, this could only be done
in multi-expansion accordions.

This is accomplished by manually adding a "None" option to the initial
open panel dropdown in the edit dialog. Additionally, the code that
automatically opens the first panel if none is set as open (for single
expansion accordions) is removed from the site clientlib such that if
no panel is specified to be open on load, then no panel will be open on
load.

The documentation is updated to clarify that no panel is _forced_ to be
open; rather, only one _can_ be opened.

----
issue #1256

Co-authored-by: Burkhard Pauli <mail@bpauli.de>