[Image V3] Hide "Disable lazy loading" if disabled by policy
Changes the V3 dialog such that the "Disable lazy loading" option
is hidden if "Disable lazy loading" is set in the policy (and theu
lazy loading does not need to be disabled on individual component
instances).

Clarifies the field description to make it clear that the checkbox
only applies to the image component instance currently being edited.

Adds `@deleteHint=true` to the "Disable lazy loading" dialog option
because `@disableLazyLoading=false` has no meaning. (same behviour
as when omitted).

----
refs #2311