Update Image V3 test data
Replace the use of `"ture"` and `"false"` in the
V3 image test data with `true` and `false` (i.e.
boolean instead of string).

This appears to be required when building in Java 17,
possibly due to an issue with the converter in Java 12+.

See FELIX-6157: https://issues.apache.org/jira/browse/FELIX-6157.