Remove `FormsHelperStubber`
`FormsHelperStubber` uses reflection to modify the
`FormsHelper` class provided by AEM. This is done
with the help of the `javassist` test dependency.

There is no (reasonable) update that could be made
to allow this to function in both Java 8 and Java 17.

To correct this issue, `FormsHelperStubber` is removed
entirely and replaced with an actual stub of the class
`day.cq.wcm.foundation.forms.FormsHelper`.