Fix project not building with Java 17 (issue #2357) (#2372)
* Update `mockito` -> 4.8.1

Update `mockito` to version 4.8.1, which is the latest version
at this time.

Update `OEmbedClientImplTest` and `ContainerServletTest`to remove
references to removed class`org.mockito.internal.util.reflection.FieldSetter`.
`org.apache.commons.lang3.reflect.FieldUtils` is used instead.

In `AmpModeForwardFilterTest`, the deprecated and removed `anyObject()`
method is removed and replaced with `ArgumentMatchers.any()`.

* Update `caconfig-mock-plugin` -> 1.4.0

Update `caconfig-mock-plugin` to version 1.4.0, which
is the latest version at this time.

Version 1.3.4 or higher is required for newer versions
of `aem-mock` (>=4.1.6).

* Upgrade `junit-jupiter` -> 5.9.1

- Update `junit-jupiter` to version 5.9.1.
- Remove any reference to junit4 in `DependencyManagement` because
the dependencies are no longer used.
- Replace `junit-jupter-api/params/engine` dependencies with
`junit-jupiter` because it is no longer required to include each
individual dependency.

* Update `aem-mock` -> 5.1.0

Update `io.wcm.testing.aem-mock.junit5` to version 5.1.0.

- `MockExternalizerFactory` is deprecated in favour of using the
  `MockExternalizer` which is part of `aem-mock` since version 4.1.0.
- Update tests to remove usage of `MockExternalizer`.
- Correct incorrect `@type` values in exporter tests (which were the
  result of a limitation in previous versions of `aem-mock`.

* Update dependencies

Update dependencies that do not require any specific code
changes to upgrade.

* Update spotbugs -> 4.7.2.1

Updates `spotbugs` plugin from version 4.0.4 to
version 4.7.2.1.

This causes many `EI_EXPOSE_REP` and `EI_EXPOSE_REP2` errors
to be spotted. These are suppressed for the time being.

* Don't use reflection to test loggers

Update tests of log messages to use a custom log
appender instead of using reflection to override
the private static logger.

* Fix JDK17 reflection error in `FormsHelperStubber`

replace use of `CtClass.toClass()` with `CtClass.toClass(Class neighbour)`
in `FormsHelperStubber`.

This is required for Java 17. For more information see discussion:
https://github.com/jboss-javassist/javassist/issues/369

* Change `ContanerServletTest` to JCR_MOCK

Change `ContainerServletTest` to use JCR_MOCK instead of OAK_MOCK.

* Don't require values to be string in `CoreResourceWrapper`

Change the value type from `String` -> `Object` in `CoreResourceWrapper`
so that it matches up with what is allowed as per the `ValueMap`
interface.

This prevents unessisary use of string values (such as `"true"` instead
of `true` or `Boolean.TRUE`).

* Remove unnecessary use of `MockSlingHttpServletRequest`

Remove use of `MockSlingHttpServletRequest` when it is unnecessary
because a `MockSlingHttpServletRequest` already exists via the
`AemContext`.

* Fix `AdaptiveImageServletMappingConfigurationConsumerTest`

Reorder service activation to ensure services are bound to component.
The requirement to do this is likely related to SLING-11668
(https://issues.apache.org/jira/browse/SLING-11668).

* Update Image V3 test data

Replace the use of `"ture"` and `"false"` in the
V3 image test data with `true` and `false` (i.e.
boolean instead of string).

This appears to be required when building in Java 17,
possibly due to an issue with the converter in Java 12+.

See FELIX-6157: https://issues.apache.org/jira/browse/FELIX-6157.

* Remove `FormsHelperStubber`

`FormsHelperStubber` uses reflection to modify the
`FormsHelper` class provided by AEM. This is done
with the help of the `javassist` test dependency.

There is no (reasonable) update that could be made
to allow this to function in both Java 8 and Java 17.

To correct this issue, `FormsHelperStubber` is removed
entirely and replaced with an actual stub of the class
`day.cq.wcm.foundation.forms.FormsHelper`.

* Add license to `logback.xml`

Adds Adobe license header to `logback.xml` as required by the
rat plugin.

* Test servlet deregistration on deactivet of configuration consumer

Adds test validating that the Servlets registered by
`AdaptiveImageServletMappingConfigurationConsumer` are removed
when the `#deactivat()` method is called.

Co-authored-by: Burkhard Pauli <mail@bpauli.de>