Update Content Fragment List model to allow "all tag" requirement
Updates the Content Fragment List component model so to add the
ability to require all tags to match instead of any tag (i.e. us
`AND` instead of the default `OR`).

This is similar to the List component and makes use of the same
property name (`./tagsMatch`).

Additionally, injected properties are removed when they can easily
be derived from other injected objects and aren't used outside of
the initializer method - and so there is no reason to hold on to them
for the entire lifespan of the object.

Furthermore, resources resolved from the leaky query builder resource
resolver are re-resolved with the non-leaky resource resolver before
being used. This is done to ensure that no reference to the leaky
resource resolver persists after it is closed.

----
refs #2381